* fixed/extended soft 128-bit integer support (+enabled these in math:: for 32-bit targets now)  * const_math: args in math forwarders are now const  * const_math/rt_math: better __uint128_t handling in clz/ctz/popcount (works with soft __uint128_t as well now)  * gl_support: added ranged buffer mapping enums for gles 2.0 (ext -> core)  * misc 32-bit target size_t casting/handling fixes  * xcode project: use armv7 instead of armv7s as a valid arch (should build for 32-bit generic ios device now + armv7 got tls support in ios 9.0 -> should actually work again)

+++ core/file_io.cpp
@@ -492,7 +492,7 @@ bool file_io::read_file(stringstream& buffer) {
 #endif
 	
 	const unsigned long long int size = (unsigned long long int)size_ll;
-	auto data = make_unique<char[]>(size + 1u);
+	auto data = make_unique<char[]>(size_t(size + 1u));
 	if(data == nullptr) return false;
 	
 	memset(data.get(), 0, size_t(size + 1u));
